refactor: move normalize_vegalite_spec into lumen.ai.utils
Following review feedback, relocate the reusable normalize_vegalite_spec
function and its private geographic helpers from the VegaLiteAgent module
into lumen.ai.utils, so spec normalization lives with the other shared AI
utilities. VegaLiteAgent._extract_spec imports it back and still delegates,
so behavior is unchanged. editor_type now defaults to None and lazily
imports VegaLiteEditor to avoid an editors/utils import cycle.
