+++ lumen/ai/agents/vega_lite.py
@@ -17,120 +17,18 @@ from ..code_executor import AltairExecutor, CodeSafetyCheck
 from ..config import (
     LUMEN_CACHE_DIR, PROMPTS_DIR, VECTOR_STORE_ASSETS_URL,
     VEGA_LITE_EXAMPLES_NUMPY_DB_FILE, VEGA_LITE_EXAMPLES_OPENAI_DB_FILE,
-    VEGA_MAP_LAYER, VEGA_ZOOMABLE_MAP_ITEMS, UserCancelledError,
+    UserCancelledError,
 )
 from ..context import TContext
 from ..editors import LumenEditor, VegaLiteEditor
 from ..llm import Message, OpenAI
 from ..models import EscapeBaseModel, RetrySpec
 from ..utils import (
-    get_data, get_schema, load_json, log_debug, retry_llm_output,
+    get_data, get_schema, load_json, log_debug, normalize_vegalite_spec,
+    retry_llm_output,
 )
 from ..vector_store import DuckDBVectorStore
 from .base_code import BaseCodeAgent
-
-
-def normalize_vegalite_spec(
-    vega_spec: dict[str, Any], *, editor_type: type[VegaLiteEditor] = VegaLiteEditor
-) -> dict[str, Any]:
-    """Normalize and validate a parsed Vega-Lite spec without an agent instance.
-
-    Injects ``$schema``, defaults ``width``/``height`` to ``"container"``
-    (skipping compound charts such as ``hconcat``/``vconcat``/``concat``/
-    ``facet``/``repeat``), validates the spec via ``editor_type.validate_spec``
-    and adds geographic pan/zoom interactivity for maps. The spec dict is
-    mutated in place and returned wrapped in the view metadata Lumen expects.
-
-    Pure helper: no LLM, no UI and no agent instance are required.
-    """
-    # Remove wrapper properties that aren't part of Vega-Lite spec
-    for key in ['sizing_mode', 'min_height', 'type']:
-        vega_spec.pop(key, None)
-
-    if "$schema" not in vega_spec:
-        vega_spec["$schema"] = "https://vega.github.io/schema/vega-lite/v5.json"
-
-    # Check if this is a compound chart (hconcat, vconcat, concat, facet, repeat)
-    is_compound = any(key in vega_spec for key in ['hconcat', 'vconcat', 'concat', 'facet', 'repeat'])
-
-    if is_compound:
-        # Remove invalid top-level width/height for compound charts
-        # Altair's config.view.continuousWidth/Height handles sub-chart sizing
-        vega_spec.pop('width', None)
-        vega_spec.pop('height', None)
-    else:
-        if "width" not in vega_spec:
-            vega_spec["width"] = "container"
-        if "height" not in vega_spec:
-            vega_spec["height"] = "container"
-
-    editor_type.validate_spec(vega_spec)
-
-    # using string comparison because these keys could be in different nested levels
-    vega_spec_str = dump_yaml(vega_spec)
-    # Handle different types of interactive controls based on chart type
-    if "latitude:" in vega_spec_str or "longitude:" in vega_spec_str:
-        vega_spec = _add_geographic_items(vega_spec, vega_spec_str)
-    # elif ("point: true" not in vega_spec_str or "params" not in vega_spec) and vega_spec_str.count("encoding:") == 1:
-    #     # add pan/zoom controls to all plots except geographic ones and points overlaid on line plots
-    #     # because those result in an blank plot without error
-    #     vega_spec["params"] = [{"bind": "scales", "name": "grid", "select": "interval"}]
-    return {"spec": vega_spec, "sizing_mode": "stretch_both", "min_height": 200}
-
-
-def _add_zoom_params(vega_spec: dict) -> None:
-    """Add zoom parameters to vega spec."""
-    if "params" not in vega_spec:
-        vega_spec["params"] = []
-
-    existing_param_names = {
-        p.get("name") for p in vega_spec["params"]
-        if isinstance(p, dict) and "name" in p
-    }
-
-    for p in VEGA_ZOOMABLE_MAP_ITEMS["params"]:
-        if p.get("name") not in existing_param_names:
-            vega_spec["params"].append(p)
-
-
-def _setup_projection(vega_spec: dict) -> None:
-    """Setup map projection settings."""
-    if "projection" not in vega_spec:
-        vega_spec["projection"] = {"type": "mercator"}
-    vega_spec["projection"].update(VEGA_ZOOMABLE_MAP_ITEMS["projection"])
-
-
-def _handle_map_compatibility(vega_spec: dict, vega_spec_str: str) -> None:
-    """Handle map projection compatibility and add geographic outlines."""
-    has_world_map = "world-110m.json" in vega_spec_str
-    uses_albers_usa = vega_spec["projection"]["type"] == "albersUsa"
-
-    if has_world_map and uses_albers_usa:
-        # albersUsa incompatible with world map
-        vega_spec["projection"] = "mercator"
-    elif not has_world_map and not uses_albers_usa:
-        # Add world map outlines if needed
-        if "layer" not in vega_spec:
-            vega_spec["layer"] = [{
-                "mark": vega_spec.pop("mark", {})
-            }]
-        vega_spec["layer"].insert(0, VEGA_MAP_LAYER["world"])
-
-
-def _add_geographic_items(vega_spec: dict, vega_spec_str: str) -> dict:
-    """Add geographic visualization items to vega spec."""
-    _add_zoom_params(vega_spec)
-    _setup_projection(vega_spec)
-
-    # Remove projection from individual layers to prevent conflicts
-    # All layers must inherit the top-level projection for zoom/pan to work correctly
-    if "layer" in vega_spec:
-        for layer in vega_spec["layer"]:
-            if isinstance(layer, dict) and "projection" in layer:
-                del layer["projection"]
-
-    _handle_map_compatibility(vega_spec, vega_spec_str)
-    return vega_spec
 
 
 class VegaLiteSpec(EscapeBaseModel):
