+++ lumen/ai/utils.py
@@ -38,18 +38,20 @@ from jsonschema import ValidationError
 from markupsafe import escape
 from panel_material_ui import Details
 
+from ..config import dump_yaml
 from ..pipeline import Pipeline
 from ..sources.base import Source
 from ..transforms import SQLRemoveSourceSeparator
 from ..util import log
 from .config import (
-    PROMPTS_DIR, SOURCE_TABLE_SEPARATOR, UNRECOVERABLE_ERRORS,
-    MissingContextError, RetriesExceededError,
+    PROMPTS_DIR, SOURCE_TABLE_SEPARATOR, UNRECOVERABLE_ERRORS, VEGA_MAP_LAYER,
+    VEGA_ZOOMABLE_MAP_ITEMS, MissingContextError, RetriesExceededError,
 )
 
 if TYPE_CHECKING:
     from panel.chat.step import ChatStep
 
+    from .editors import VegaLiteEditor
     from .models import (
         DeleteLine, InsertLine, LineEdit, ReplaceLine,
     )
@@ -1485,3 +1487,112 @@ def result_to_dataframe(result) -> pd.DataFrame | None:
             return None
 
     return None
+
+
+def normalize_vegalite_spec(
+    vega_spec: dict[str, Any], *, editor_type: type[VegaLiteEditor] | None = None
+) -> dict[str, Any]:
+    """Normalize and validate a parsed Vega-Lite spec without an agent instance.
+
+    Injects ``$schema``, defaults ``width``/``height`` to ``"container"``
+    (skipping compound charts such as ``hconcat``/``vconcat``/``concat``/
+    ``facet``/``repeat``), validates the spec via ``editor_type.validate_spec``
+    and adds geographic pan/zoom interactivity for maps. The spec dict is
+    mutated in place and returned wrapped in the view metadata Lumen expects.
+
+    Pure helper: no LLM, no UI and no agent instance are required. ``editor_type``
+    defaults to ``VegaLiteEditor`` when not provided.
+    """
+    if editor_type is None:
+        # Imported lazily to avoid an editors -> utils import cycle.
+        from .editors import VegaLiteEditor
+        editor_type = VegaLiteEditor
+
+    # Remove wrapper properties that aren't part of Vega-Lite spec
+    for key in ['sizing_mode', 'min_height', 'type']:
+        vega_spec.pop(key, None)
+
+    if "$schema" not in vega_spec:
+        vega_spec["$schema"] = "https://vega.github.io/schema/vega-lite/v5.json"
+
+    # Check if this is a compound chart (hconcat, vconcat, concat, facet, repeat)
+    is_compound = any(key in vega_spec for key in ['hconcat', 'vconcat', 'concat', 'facet', 'repeat'])
+
+    if is_compound:
+        # Remove invalid top-level width/height for compound charts
+        # Altair's config.view.continuousWidth/Height handles sub-chart sizing
+        vega_spec.pop('width', None)
+        vega_spec.pop('height', None)
+    else:
+        if "width" not in vega_spec:
+            vega_spec["width"] = "container"
+        if "height" not in vega_spec:
+            vega_spec["height"] = "container"
+
+    editor_type.validate_spec(vega_spec)
+
+    # using string comparison because these keys could be in different nested levels
+    vega_spec_str = dump_yaml(vega_spec)
+    # Handle different types of interactive controls based on chart type
+    if "latitude:" in vega_spec_str or "longitude:" in vega_spec_str:
+        vega_spec = _add_geographic_items(vega_spec, vega_spec_str)
+    # elif ("point: true" not in vega_spec_str or "params" not in vega_spec) and vega_spec_str.count("encoding:") == 1:
+    #     # add pan/zoom controls to all plots except geographic ones and points overlaid on line plots
+    #     # because those result in an blank plot without error
+    #     vega_spec["params"] = [{"bind": "scales", "name": "grid", "select": "interval"}]
+    return {"spec": vega_spec, "sizing_mode": "stretch_both", "min_height": 200}
+
+
+def _add_zoom_params(vega_spec: dict) -> None:
+    """Add zoom parameters to vega spec."""
+    if "params" not in vega_spec:
+        vega_spec["params"] = []
+
+    existing_param_names = {
+        p.get("name") for p in vega_spec["params"]
+        if isinstance(p, dict) and "name" in p
+    }
+
+    for p in VEGA_ZOOMABLE_MAP_ITEMS["params"]:
+        if p.get("name") not in existing_param_names:
+            vega_spec["params"].append(p)
+
+
+def _setup_projection(vega_spec: dict) -> None:
+    """Setup map projection settings."""
+    if "projection" not in vega_spec:
+        vega_spec["projection"] = {"type": "mercator"}
+    vega_spec["projection"].update(VEGA_ZOOMABLE_MAP_ITEMS["projection"])
+
+
+def _handle_map_compatibility(vega_spec: dict, vega_spec_str: str) -> None:
+    """Handle map projection compatibility and add geographic outlines."""
+    has_world_map = "world-110m.json" in vega_spec_str
+    uses_albers_usa = vega_spec["projection"]["type"] == "albersUsa"
+
+    if has_world_map and uses_albers_usa:
+        # albersUsa incompatible with world map
+        vega_spec["projection"] = "mercator"
+    elif not has_world_map and not uses_albers_usa:
+        # Add world map outlines if needed
+        if "layer" not in vega_spec:
+            vega_spec["layer"] = [{
+                "mark": vega_spec.pop("mark", {})
+            }]
+        vega_spec["layer"].insert(0, VEGA_MAP_LAYER["world"])
+
+
+def _add_geographic_items(vega_spec: dict, vega_spec_str: str) -> dict:
+    """Add geographic visualization items to vega spec."""
+    _add_zoom_params(vega_spec)
+    _setup_projection(vega_spec)
+
+    # Remove projection from individual layers to prevent conflicts
+    # All layers must inherit the top-level projection for zoom/pan to work correctly
+    if "layer" in vega_spec:
+        for layer in vega_spec["layer"]:
+            if isinstance(layer, dict) and "projection" in layer:
+                del layer["projection"]
+
+    _handle_map_compatibility(vega_spec, vega_spec_str)
+    return vega_spec
